django/requirements.txt
django
djangorestframework
gunicorn
psycopg2 